Add document touch listeners lazily
Only add document touch listeners when anyone is actually interested in them.

As `updatePointsLocation()` replaces the content of  `points` every time it's invoked there's really no need to track anything prior to the creation of a multitouch input.

(Only small edge case is if a touch is already happening when the multitouch input is happening – then that touch won't be detected until the next `touchstart` or `touchmove` – but I think that is acceptable considering the benefits of not _always_ having touch listeners when one includes Popmotion, no matter if one uses the multitouch input or not?)